Use uint64 signatures

+++ src/ducc0/wgridder/wgridder.h
@@ -39,8 +39,8 @@ template<typename Tcalc, typename Tacc, typename Tms, typename Tms_in=cmav<compl
 template<typename Tcalc, typename Tacc, typename Tms, typename Tms_in=cmav<complex<Tms>,1>, typename Timg>
   void ms2dirty_bda(
     const cmav<double,2> &uvw,                        // (nrows,3)
-    const cmav<size_t,1> &freqlist_id,                // (nrows),
-    const cmav<size_t,1> &freqlist_nfreqs,            // (max(freqlist_id)+1)
+    const cmav<uint64_t,1> &freqlist_id,              // (nrows),
+    const cmav<uint64_t,1> &freqlist_nfreqs,          // (max(freqlist_id)+1)
     const cmav<double,1> &freqlist_freqs,             // (sum(freqlist_nfreqs), concatenated frequency lists for all freqlist_ids
     const Tms_in &ms,                                 // concatenated array of visibilities, shape (sum_i(freqlist_nfreq[freqlist_id[i]))
     const cmav<Tms,1> &wgt_,                          // same shape as above
@@ -60,8 +60,8 @@ template<typename Tcalc, typename Tacc, typename Tms, typename Timg> void dirty2
 template<typename Tcalc, typename Tacc, typename Tms, typename Timg>
   void dirty2ms_bda(
     const cmav<double,2> &uvw,
-    const cmav<size_t,1> &freqlist_id,                // (nrows),
-    const cmav<size_t,1> &freqlist_nfreqs,            // (max(freqlist_id)+1)
+    const cmav<uint64_t,1> &freqlist_id,              // (nrows),
+    const cmav<uint64_t,1> &freqlist_nfreqs,          // (max(freqlist_id)+1)
     const cmav<double,1> &freqlist_freqs,             // (sum(freqlist_nfreqs), concatenated frequency lists for all freqlist_ids
     const cmav<Timg,2> &dirty,
     const cmav<Tms,1> &wgt_, const cmav<uint8_t,1> &mask_, double pixsize_x, double pixsize_y,
